requirements.txt
fastapi
uvicorn
openai
psycopg2
numpy
scipy
python-dotenv